back/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
Flask==3.0.3
SQLAlchemy==2.0.31
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.1
matplotlib==3.10.3